refactor(auth): same schema style as recipes
Refs: #101
