+++ apps/backend/src/modules/auth/auth.service.ts
@@ -1,4 +1,4 @@
-import type { AuthResponse, LoginBody, RegisterBody } from "@recipes/shared";
+import type { AuthResponse, LoginInput, RegisterInput } from "@recipes/shared";
 import { ConflictError, UnauthorizedError } from "@/common/errors.js";
 import type { Logger } from "@/common/logger.js";
 import type { PasswordService } from "@/common/passwords/password.service.js";
@@ -7,8 +7,8 @@ import { toUser } from "@/modules/users/user.mapper.js";
 import type { UserRepository } from "@/modules/users/user.repository.js";
 
 export interface AuthService {
-  register(data: RegisterBody): Promise<AuthResponse>;
-  login(data: LoginBody): Promise<AuthResponse>;
+  register(data: RegisterInput): Promise<AuthResponse>;
+  login(data: LoginInput): Promise<AuthResponse>;
 }
 
 type UserRepositoryPort = Pick<UserRepository, "findOne" | "exists" | "create">;
