+++ packages/shared/src/auth/auth.response.schema.ts
@@ -1,0 +1,9 @@
+import { z } from "zod";
+import { userSchema } from "../users/user.schema.js";
+
+export const authResponseSchema = z.object({
+  user: userSchema,
+  token: z.string(),
+});
+
+export type AuthResponse = z.infer<typeof authResponseSchema>;
